Iterate over <strong> tags to look for source file.
This makes the search for latest source tarball version more robust,
e.g. in the case of rivet, for which the 2nd <strong> element does
not refer to Rivet tarballs, but to some example code.

If an extension string is specified duriing the hepVer() call, the
search looks for a matching source tarball, otherwise it looks for
a tarball with a few common extensions (.gz, .bz2, etc.)