fix: Print correct pkg when GH prj is not found.
In this case, the `ghrepo` may not be the same as the pkg name, in case
the it is determined to be an unresolved macro. So we raise an exception
from ghLastVer() and actually append to the list of errors from the main
func where pkg is still well defined.