fix: Resolve custom rpm macro in GH project/repo.
If the src url references a macro for the GH project/repo, grep through
the spec file to find a definition for the macro and use its value. This
is useful, in particular, for multi-flavour packages where a custom
macro, say '%pname' may be used instead of the bare pkg name in the URL.
Note that, in a majority of these cases, `rpmspec` would not work
either, so manual grep-ing is our best bet.