fix: Simplify appname stripping from tag name.
Instead of a generic regex match to determine where the app name ends
and versioning begins, let the base scrapers return the full tag name
back without parsing it through `stdver`. Do this parsing at the level
of the derived feed reader where the appname is known, and be specific
to this appname when stripping from tag.