fix: Improve handling of permanent redirects.
For HTTP 301/308, update feed URL (which is eventually saved to cache
file) to the redirected destination, so that for subsequent attempts,
when we use the feed_url from the cache, we directly target the new
destination instead of repeatedly hitting 301/308.

Also handle the case where 301/308 redirected feed may be empty because
of matching 'etag', indicating no update of feed data is needed.