refactor: PyPI: Allow getting prj name from urls.
Refactor PyPIVersion (previously PyPI):
* Change class init to accept pkg_cache_dir, src_url, obs_prj name, and
  url (optional) as arguments.
* If the optional url is passed, it must refer to the PyPI project url
  and is used directly to determine the project name.
* Otherwise, the first attempt to set project name is made using
  src_url, which must refer to "*.pythonhosted.org".
* Finally, the obs_prj name, sans a leading "python[23]?-" is used as
  the project name.

If there is a pypi cache file, short-circuit all this and directly use
the value of "project" from the cache.

All this to allow the use of PyPIVersion in cases where neither the url,
src_url, not obs prj names provide the correct hints to the upstream
PyPI project. See, for example, the case of
obs://science/python-Mathics (PyPI: "Mathics3").